price-info does not depend on util
