+++ price-info/Cargo.toml
@@ -13,4 +13,4 @@ log = "0.3"
 serde_json = "1.0"
 
 [dev-dependencies]
-ethcore-util = { path = "../util" }
+parking_lot = "0.4"
